#53 moving RSA types in to their own folders
