+++ cas-dotnet-sdk-tests/RustRsaWrapperTests.cs
@@ -1,4 +1,5 @@
 ﻿using CasDotnetSdk.Asymmetric;
+using CasDotnetSdk.Asymmetric.Types;
 using System.Text;
 using Xunit;
 using static CasDotnetSdk.Asymmetric.RSAWrapper;
