+++ cas-dotnet-sdk/Asymmetric/Types/RsaDecryptBytesResult.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace CasDotnetSdk.Asymmetric.Types
+{
+    internal struct RsaDecryptBytesResult
+    {
+        public IntPtr decrypted_result_ptr;
+        public int length;
+    }
+}
