+++ cas-dotnet-sdk/Asymmetric/Types/RsaEncryptBytesResult.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace CasDotnetSdk.Asymmetric.Types
+{
+    internal struct RsaEncryptBytesResult
+    {
+        public IntPtr encrypted_result_ptr;
+        public int length;
+    }
+}
