+++ cas-dotnet-sdk/Asymmetric/Types/RsaKeyPairResult.cs
@@ -1,0 +1,8 @@
+﻿namespace CasDotnetSdk.Asymmetric.Types
+{
+    public class RsaKeyPairResult
+    {
+        public string PublicKey { get; set; }
+        public string PrivateKey { get; set; }
+    }
+}
