+++ cas-dotnet-sdk/Asymmetric/Types/RsaKeyPairStruct.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace CasDotnetSdk.Asymmetric.Types
+{
+    internal struct RsaKeyPairStruct
+    {
+        public IntPtr pub_key;
+        public IntPtr priv_key;
+    }
+}
