+++ cas-dotnet-sdk/Asymmetric/Types/RsaSignBytesResults.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace CasDotnetSdk.Asymmetric.Types
+{
+    internal struct RsaSignBytesResults
+    {
+        public IntPtr signature_raw_ptr;
+        public int length;
+    }
+}
