+++ cas-dotnet-sdk/Asymmetric/Types/RsaSignResult.cs
@@ -1,0 +1,8 @@
+﻿namespace CasDotnetSdk.Asymmetric.Types
+{
+    public class RsaSignResult
+    {
+        public string Signature { get; set; }
+        public string PublicKey { get; set; }
+    }
+}
