+++ cas-dotnet-sdk/Asymmetric/Types/RsaSignResultStruct.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace CasDotnetSdk.Asymmetric.Types
+{
+    internal struct RsaSignResultStruct
+    {
+        public IntPtr signature;
+        public IntPtr public_key;
+    }
+}
